requirements.txt
django
djangorestframework
redis
apscheduler
requests
pandas